+++ src/main/java/chess/SparkChessApplication.java
@@ -1,22 +1,16 @@
 package chess;
 
-import spark.ModelAndView;
-import spark.template.handlebars.HandlebarsTemplateEngine;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import static spark.Spark.get;
+import static spark.Spark.port;
 
 public class SparkChessApplication {
+/*
     public static void main(String[] args) {
-        get("/", (req, res) -> {
-            Map<String, Object> model = new HashMap<>();
-            return render(model, "index.hbs");
-        });
-    }
+        port(8082);
+        staticFiles.location("/static");
 
-    private static String render(Map<String, Object> model, String templatePath) {
-        return new HandlebarsTemplateEngine().render(new ModelAndView(model, templatePath));
+        Connector connector = new Connector();
+        WebChessController webChessController = new WebChessController(new ChessService(new GameDao(connector), new BoardDao(connector)));
+        webChessController.run();
     }
+*/
 }
